Back-End/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
langchain==0.3.10
langchain-openai==0.2.10
langchain-community==0.3.10
langchain-core==0.3.22
openai==1.57.2
chromadb==0.5.23
pypdf==5.1.0
tiktoken==0.8.0
